refactor(web): client side js
