+++ zellij-client/assets/utils.js
@@ -1,0 +1,11 @@
+/**
+ * Utility functions for the terminal web client
+ */
+
+/**
+ * Check if the current page is served over HTTPS
+ * @returns {boolean} true if protocol is https:, false otherwise
+ */
+export function is_https() {
+    return document.location.protocol === "https:";
+}
